门禁错误修改

+++ interfaces/inner_api/app_manager/src/appmgr/app_mgr_proxy.cpp
@@ -282,8 +282,7 @@ int32_t AppMgrProxy::NotifyMemoryLevel(int32_t level)
         return ERR_NULL_OBJECT;
     }
     int32_t ret =
-        remote->SendRequest(static_cast<uint32_t>(IAppMgr::Message::APP_NOTIFY_MEMORY_LEVEL),
-        data, reply, option);
+        remote->SendRequest(IAppMgr::Message::APP_NOTIFY_MEMORY_LEVEL, data, reply, option);
     if (ret != NO_ERROR) {
         HILOG_WARN("SendRequest is failed, error code: %{public}d", ret);
     }
